Game Objects page tutorial setup from home
